Added new pages
