+++ src/App.jsx
@@ -1,14 +1,15 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import axios from "axios";
 import { useState, useEffect } from "react";
-import MainPage from "./pages/MainPage";
+import AboutPage from "./pages/AboutPage";
 import './App.css';
 import BlogsPage from "./pages/BlogsPage";
 import ProjectsPage from "./pages/ProjectsPage";
 import LandingPanel from "./pages/LandingPanel";
+import Connect from "./pages/Connect";
 
 function App() {
-  const [isLightMode, setIsLightMode] = useState(true);
+  const [isLightMode, setIsLightMode] = useState(false);
   const [blogsArray, setBlogsArray] = useState([]);
 
   useEffect(() => {
@@ -26,10 +27,11 @@ function App() {
   return (
     <BrowserRouter>
       <Routes>
-        <Route path="/" element={<MainPage isLightMode={isLightMode} setIsLightMode={setIsLightMode} />} />
-        <Route path="/blogs" element={<BlogsPage blogsArray={blogsArray} />} />
-        <Route path="/projects" element={<ProjectsPage />} />
-        <Route path="/panel" element={<LandingPanel />} />
+        <Route path="/" element={<LandingPanel />} />
+        <Route path="/aboutPage" element={<AboutPage isLightMode={isLightMode} setIsLightMode={setIsLightMode} />} />
+        <Route path="/blogs" element={<BlogsPage blogsArray={blogsArray} isLightMode={isLightMode} setIsLightMode={setIsLightMode} />} />
+        <Route path="/projects" element={<ProjectsPage isLightMode={isLightMode} setIsLightMode={setIsLightMode} />} />
+        <Route path="/connect" element={<Connect />} />
 
       </Routes>
     </BrowserRouter>
