+++ src/pages/LandingPanel.jsx
@@ -1,38 +1,52 @@
+import React from "react";
+import { useNavigate } from "react-router-dom";
+import logo from "../assets/logo.svg";
+import projectsIcon from "../assets/projects.svg";
+import blogsIcon from "../assets/blogs.svg";
+import socialsIcon from "../assets/socials.svg";
+
+const TableOfContentItem = ({ text, imgSrc, onClick }) => (
+    <div className="table-of-contents-grid" onClick={onClick} role="button" tabIndex={0}>
+        <span>{text}</span>
+        <img src={imgSrc} alt={text} />
+    </div>
+);
+
 const LandingPanel = () => {
+    const navigate = useNavigate();
+    const handleNavigation = (route) => {
+        navigate(route);
+    }
+
     return (
         <div className="page" id="landing-panel">
             <section className="hero">
                 <h1>Elhussin Farah</h1>
             </section>
 
-            {/*   <img src="src/assets/logo.svg" /> */}
-
             <section className="table-of-contents">
                 <h1>Table of Content</h1>
                 <nav id="table-content-grid">
-                    <div className="table-of-contents-grid">
-                        <a href="#about-me">#1 About Me</a>
-                        <img src="src/assets/logo.svg" />
-                    </div>
-
-                    <div className="table-of-contents-grid">
-                        <a href="#projects">#2 Projects</a>
-                        <img src="src/assets/projects.svg" />
-
-                    </div>
-                    <div className="table-of-contents-grid">
-                        <a href="#blog">#3 Blog</a>
-                        <img src="src/assets/blogs.svg" />
-                    </div>
-                    <div className="table-of-contents-grid">
-                        <a href="#connect">#4 Connect</a>
-                        <img src="src/assets/socials.svg" />
-
-                    </div>
-
-
-
-
+                    <TableOfContentItem
+                        text="#1 About Me"
+                        imgSrc={logo}
+                        onClick={() => handleNavigation("/aboutPage")}
+                    />
+                    <TableOfContentItem
+                        text="#2 Projects"
+                        imgSrc={projectsIcon}
+                        onClick={() => handleNavigation("/projects")}
+                    />
+                    <TableOfContentItem
+                        text="#3 Blog"
+                        imgSrc={blogsIcon}
+                        onClick={() => handleNavigation("/blogs")}
+                    />
+                    <TableOfContentItem
+                        text="#4 Connect"
+                        imgSrc={socialsIcon}
+                        onClick={() => handleNavigation("/connect")}
+                    />
                 </nav>
             </section>
         </div>
