+++ src/pages/ProjectsPage.jsx
@@ -1,11 +1,17 @@
 import { useEffect, useState } from "react";
-
-function ProjectsPage() {
+import { useNavigate } from "react-router-dom";
+import moonIcon from '../assets/dark.svg';
+import sunIcon from '../assets/light.svg';
+function ProjectsPage(isLightMode, setIsLightMode) {
     const [isOverlayVisible, setOverlayVisible] = useState(false);
     const [currentImages, setCurrentImages] = useState([]);
 
+    const navigate = useNavigate();
 
 
+    const handleNavigation = (route) => {
+        navigate(route);
+    }
 
     const showOverlay = (images) => {
         const element = document.querySelector(".overlay")
@@ -54,38 +60,88 @@ function ProjectsPage() {
 
     return (
         <div id="projects-page">
+            <nav className='nav'>
+                <h1>
+                    Projects
+                </h1>
+                <div className="table-of-contents-grid">
+                    <a
+                        onClick={() => handleNavigation("/")}
+                    >Home</a>
+
+
+                </div>
+                {/* <div className="table-of-contents-grid">
+                            <a href="#connect">Home</a>
+
+
+                        </div> */}
+
+                <div className="table-of-contents-grid" onClick={() => { setIsLightMode((prev) => { return !prev; }) }} >
+
+                    <img
+                        src={!isLightMode ? moonIcon : sunIcon}
+                        alt={!isLightMode ? 'Dark Mode' : 'Light Mode'}
+                        className={`theme-icon ${!isLightMode ? 'rotate-icon' : ''}`}
+                    />
+
+
+                </div>
+
+
+
+                <div className="table-of-contents-grid">
+                    <a
+
+                        onClick={() => handleNavigation("/blogs")}
+                    > Blogs</a>
+
+
+                </div>
+            </nav>
             {projects.map((project, index) => (
                 <div key={index} className="project-container">
                     {/*  <div className="project-container-content"> */}
                     <h1>{project.title}</h1>
-                    <div className="project-container-body">
-                        <div className="categories">
-                            {project.techStack.map((category, catIndex) => (
-                                <span key={catIndex} className="category-area">
-                                    {category}
-                                </span>
-                            ))}
-                        </div>
-                        <div
-                            style={{
-                                height: "100%",
-                                padding: "5%",
-                                display: "flex",
-                                flexDirection: "column",
-                                justifyContent: "space-evenly"
-                            }}
-                        >
-                            <p style={{ marginBottom: "5%" }}>{project.description}</p>
-                            <div className="custom-shadow-button">
-                                <button onClick={() => window.open(project.githubRepo)}><span>Repo</span></button>
-                            </div>
-                            <br /><br />
-                            <div className="custom-shadow-button">
-                                <button onClick={() => showOverlay(project.images)}><span>Gallery</span></button>
-                            </div>
-                        </div>
+                    {/* <div className="project-container-body"> */}
+                    <div className="categories">
+                        {project.techStack.map((category, catIndex) => (
+                            <span key={catIndex} className="category-area">
+                                {category}
+                            </span>
+                        ))}
                     </div>
 
+                    <p >{project.description}</p>
+                    <div className="projects-links">
+                        <div style={{ width: "fit-content" }} className="table-of-contents-grid">
+                            <a
+                                onClick={() => window.open(project.githubRepo)}
+
+                            >Repo</a>
+
+
+                        </div>
+                        <div style={{ width: "fit-content" }} className="table-of-contents-grid">
+                            <a
+                                onClick={() => showOverlay(project.images)}
+
+                            > Gallery </a>
+
+
+                        </div>
+                    </div>
+                    {/* <div className="custom-shadow-button">
+                            <button onClick={() => window.open(project.githubRepo)}><span>Repo</span></button>
+                        </div>
+                        <br /><br />
+                        <div className="custom-shadow-button">
+                            <button onClick={() => showOverlay(project.images)}><span>Gallery</span></button>
+                        </div> */}
+
+                    {/*  </div> */}
+                    <br />
+                    <br />
                 </div>
             ))}
